+++ Magic/src/resource-pack/build.sh
@@ -12,7 +12,7 @@ mkdir default
 cd default
 cp -R ../../default/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-RP-6-4.zip *
+zip -r -X ../Magic-RP-7.zip *
 cd ..
 
 echo "** BUILDING DEFAULT-Skulls **"
@@ -32,7 +32,7 @@ rm -R assets/minecraft/models/item/brushes
 rm assets/minecraft/models/item/diamond_axe.json
 rm assets/minecraft/models/item/diamond_hoe.json
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-skulls-RP-6-4.zip *
+zip -r -X ../Magic-skulls-RP-7.zip *
 cd ..
 
 echo "** BUILDING SKULLS **"
@@ -52,7 +52,7 @@ cd painterly
 cp -R ../../default/* .
 cp -R ../../painterly/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-painterly-RP-6-4.zip *
+zip -r -X ../Magic-painterly-RP-7.zip *
 cd ..
 
 echo "** BUILDING POTTER **"
@@ -62,7 +62,7 @@ cd potter
 cp -R ../../default/* .
 cp -R ../../potter/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-potter-RP-6-4.zip *
+zip -r -X ../Magic-potter-RP-7.zip *
 cd ..
 
 echo "** BUILDING WAR **"